--- src/main/java/by/owm/domain/jpa/TokenRepository.java
@@ -1,7 +1,0 @@
-package by.owm.domain.jpa;
-
-/**
- * Created by haria on 11.10.17.
- */
-public interface TokenRepository {
-}
