+++ src/main/java/by/owm/domain/repository/RolesRepository.java
@@ -1,14 +1,13 @@
-package by.owm.domain.jpa;
+package by.owm.domain.repository;
 
 import by.owm.domain.entity.RoleEntity;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
 
 import javax.validation.constraints.NotNull;
 
-/**
- * Created by haria on 13.10.17.
- */
+@Repository
 public interface RolesRepository extends MongoRepository<RoleEntity, String> {
-
-    public RoleEntity findByAuthority(@NotNull final String authority);
+    
+    RoleEntity findByAuthority(@NotNull final String authority);
 }
