+++ src/main/java/by/owm/domain/repository/UserRepository.java
@@ -1,17 +1,15 @@
-package by.owm.domain.jpa;
+package by.owm.domain.repository;
 
 import by.owm.domain.entity.UserEntity;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-/**
- * Created by haria on 10.10.17.
- */
 @Repository
 public interface UserRepository extends MongoRepository<UserEntity, String> {
     /**
      * Finds user by his email and password.
-     * @param email - users email used as login.
+     *
+     * @param email    - users email used as login.
      * @param password - users password, should be encoded.
      * @return UserEntity.
      */
@@ -19,11 +17,11 @@ public interface UserRepository extends MongoRepository<UserEntity, String> {
 
     /**
      * Finds user by token generated on registration.
+     *
      * @param token - unic token that contains encoded information about user.
      * @return unic token.
      */
     public UserEntity findUserByToken(final String token);
 
 //    public boolean updateUser(final UserEntity userEntity);
-
 }
